backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy==2.0.36
pydantic==2.10.3
pydantic-settings==2.6.1
python-multipart==0.0.17
jinja2==3.1.4
pyppeteer==2.0.0
python-dotenv==1.0.1